Fix transversal mode login
Login methods were using the login name as the key for procedures
This was not robust as
1. login name is not enforced UNIQUE in database
2. we authenticated based on the user ID and loaded user on login

This fix is kludgy as hell but is at least a step in the right direction.
Now we use the ID if we have it.
This shouldn't break any legacy authentication methods as I kept using login if no ID is
retrieved.